+++ src/store/blocks/api.ts
@@ -1,6 +1,6 @@
 import { Block, BlockSaveBase, MetricsResult } from '../state';
 
-import { get, post } from './fetch';
+import { get, post } from '../../../core/fetch';
 
 export function fetchBlock(id: string): Promise<Block> {
   return get(`/blocks/${encodeURIComponent(id)}`);
